# server/openapi/openapi.yaml
openapi: 3.0.0
info:
    version: 1.0.0
    title: web-novel
    description: Tài liệu REST API web đọc truyện chữ - Nhóm 18 - Simple Poll
...
paths:
...
    /api/novel/{novelId}/chapter/{chapterId}:
        get:
            tags:
                - "chapter"
            summary: Lấy thông tin chương theo Id
            security:
                - bearerAuth: []
            parameters:
                - in: path
                  name: novelId
                  required: true
                  schema:
                      type: integer
                - in: path
                  name: chapterId
                  required: true
                  schema:
                      type: integer
            responses:
                200:
                    description: OK
                    content:
                        application/json:
                            schema:
                                type: object
                                properties:
                                    chapter:
                                        type: object
                                        properties:
                                            novelId:
                                                type: integer
                                            chapterId:
                                                type: integer
                                            title:
                                                type: string
                                            chapterOrder:
                                                type: integer
                                            content:
                                                type: string
                404:
                    description: Not Found
                    content:
                        application/json:
                            schema:
                                type: object
                                properties:
                                    message:
                                        type: string
                            examples:
                                error1:
                                    summary: Không tìm thấy novelId
                                    value:
                                        message: novelId ${novelId} not found
                                error2:
                                    summary: Không tìm thấy chapterId
                                    value:
                                        message: chapterId ${chapterId} not found
        put:
            tags:
                - "chapter"
            summary: Cập nhật chương
            security:
                - bearerAuth: []
            parameters:
                - in: path
                  name: novelId
                  required: true
                  schema:
                      type: integer
                - in: path
                  name: chapterId
                  required: true
                  schema:
                      type: integer
            requestBody:
                content:
                    application/json:
                        schema:
                            type: object
                            properties:
                                chapterOrder:
                                    type: integer
                                title:
                                    type: string
                                content:
                                    type: string
            responses:
                200:
                    description: OK
                    content:
                        application/json:
                            schema:
                                type: object
                                properties:
                                    updatedChapter:
                                        type: object
                400:
                    description: Nhập thiếu trường
                    content:
                        application/json:
                            schema:
                                type: object
                                properties:
                                    message:
                                        type: string
                                        example: some fields are missing
                401:
                    description: Unauthorized
                    content:
                        application/json:
                            schema:
                                type: object
                                properties:
                                    message:
                                        type: string
                            examples:
                                error1:
                                    summary: Thiếu token
                                    value:
                                        message: access token not found
                                error2:
                                    summary: Tài khoản không có quyền tác giả
                                    value:
                                        message: this user has no author permissions
                                error3:
                                    summary: Tài khoản không có quyền sửa truyện
                                    value:
                                        message: this user has no permission to modify this novel
                404:
                    description: Not Found
                    content:
                        application/json:
                            schema:
                                type: object
                                properties:
                                    message:
                                        type: string
                            examples:
                                error1:
                                    summary: Không tìm thấy novelId
                                    value:
                                        message: novelId ${novelId} not found
                                error2:
                                    summary: Không tìm thấy chapterId
                                    value:
                                        message: chapterId ${chapterId} not found
        delete:
            tags:
                - "chapter"
            summary: Xóa chương
            security:
                - bearerAuth: []
            parameters:
                - in: path
                  name: novelId
                  required: true
                  schema:
                      type: integer
                - in: path
                  name: chapterId
                  required: true
                  schema:
                      type: integer
            responses:
                200:
                    description: OK
                    content:
                        application/json:
                            schema:
                                type: object
                                properties:
                                    deletedChapter:
                                        type: object
                400:
                    description: Nhập thiếu trường
                    content:
                        application/json:
                            schema:
                                type: object
                                properties:
                                    message:
                                        type: string
                                        example: some fields are missing
                401:
                    description: Unauthorized
                    content:
                        application/json:
                            schema:
                                type: object
                                properties:
                                    message:
                                        type: string
                            examples:
                                error1:
                                    summary: Thiếu token
                                    value:
                                        message: access token not found
                                error2:
                                    summary: Tài khoản không có quyền tác giả
                                    value:
                                        message: this user has no author permissions
                                error3:
                                    summary: Tài khoản không có quyền sửa truyện
                                    value:
                                        message: this user has no permission to modify this novel
                404:
                    description: Not Found
                    content:
                        application/json:
                            schema:
                                type: object
                                properties:
                                    message:
                                        type: string
                            examples:
                                error1:
                                    summary: Không tìm thấy novelId
                                    value:
                                        message: novelId ${novelId} not found
                                error2:
                                    summary: Không tìm thấy chapterId
                                    value:
                                        message: chapterId ${chapterId} not found
...
components:
    securitySchemes:
        bearerAuth:
            type: apiKey
            name: JWT-Authorization
            in: header